Create dictionary_attack.py
This function simulates a dictionary attack by comparing the hashed form of a user’s password with hashes of commonly used passwords stored in a file. It measures the time taken to find a match, demonstrating how easily weak and common passwords can be compromised. The attack is limited to a predefined wordlist and is implemented strictly for educational purposes.